Use separate directory for ButtonSolidLink
