+++ cardigan/stories/components/ButtonSolid.js
@@ -1,6 +1,6 @@
 import { storiesOf } from '@storybook/react';
 import ButtonSolid from '../../../common/views/components/ButtonSolid/ButtonSolid';
-import ButtonSolidLink from '../../../common/views/components/ButtonSolid/ButtonSolidLink';
+import ButtonSolidLink from '../../../common/views/components/ButtonSolidLink/ButtonSolidLink';
 import Readme from '../../../common/views/components/ButtonSolid/README.md';
 import { boolean, text } from '@storybook/addon-knobs/react';
